Update README.md
The above agreement has been removed and cancelled due to one or more of the people who signed to take back the contract, the third signer has refused to declare the rights to the other two. But as the majority have decided otherwise, this has been cancelled, Do NOT WHATSOEVER re-add the contract and false-sign the contract or you will be subject to face the law.